Move MigrationManager to separate file and run tests
