+++ src/code/migrationManager.js
@@ -1,0 +1,25 @@
+export default function MigrationManager(migrations) {
+	function getVersion(settings) {
+		if (!settings.hasOwnProperty('version')) {
+			return 0
+		}
+		return settings.version
+	}
+
+	function isMigrationNeeded(startingVersion) {
+		return startingVersion < Object.keys(migrations).pop()
+	}
+
+	this.migrate = function(settings) {
+		const startingVersion = getVersion(settings)
+		if (isMigrationNeeded(startingVersion)) {
+			for (const key in migrations) {
+				const toVersion = Number(key)
+				if (toVersion > startingVersion) {
+					migrations[toVersion](settings)
+					settings.version = toVersion
+				}
+			}
+		}
+	}
+}
